refactor: move ts file into components
