--- sass/_icons-material-design.scss
@@ -1,6 +1,0 @@
-/* This is needed for some mobile phones to display the Google Icon font properly */
-.material-icons, .material-symbols-outlined,
-.material-symbols-rounded, .material-symbols-sharp {
-  text-rendering: optimizeLegibility;
-  font-feature-settings: 'liga';
-}
