+++ sass/_table_of_contents.scss
@@ -1,15 +1,12 @@
 .table-of-contents {
   list-style: none;
-
   &.fixed {
     position: fixed;
   }
-
   li {
     padding: 0;
   }
-
-  a {
+  a { 
     display: inline-block;
     font-weight: 400;
     color: var(--md-sys-color-secondary);
@@ -17,7 +14,6 @@
     height: 2rem;
     line-height: 2rem;
     border-left: 1px solid var(--md-sys-color-outline-variant);
-    
     &:hover {
       color: var(--md-sys-color-on-background);
       padding-left: 15px;
