+++ sass/materialize.scss
@@ -10,7 +10,6 @@
 @forward "./color-variables";
 @forward "./variables";
 @forward "./global";
-@forward "./icons-material-design";
 @forward "./grid";
 // animations
 @forward "./transitions.css";
