+++ src/index.ts
@@ -15,7 +15,7 @@ import { Parallax, ParallaxOptions } from './parallax';
 import { Pushpin, PushpinOptions } from './pushpin';
 import { ScrollSpy, ScrollSpyOptions } from './scrollspy';
 import { FormSelect, FormSelectOptions } from './select';
-import { Sidenav, SidenavOptions } from './sidenav';
+import { Sidenav, SidenavOptions } from '../components/navigation-drawer/sidenav';
 import { Slider, SliderOptions } from './slider';
 import { Tabs, TabsOptions } from './tabs';
 import { TapTarget, TapTargetOptions } from './tapTarget';
@@ -44,7 +44,7 @@ export { Parallax } from './parallax';
 export { Pushpin } from './pushpin';
 export { ScrollSpy } from './scrollspy';
 export { FormSelect } from './select';
-export { Sidenav } from './sidenav';
+export { Sidenav } from '../components/navigation-drawer/sidenav';
 export { Slider } from './slider';
 export { Tabs } from './tabs';
 export { TapTarget } from './tapTarget';
